requirements.txt
yfinance
pandas
ta
gspread
oauth2client
gspread-dataframe
python-dotenv
scikit-learn 